requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
passlib[argon2]
python-jose
python-multipart
flask
requests
python-dotenv